requirements.txt
opencv-python-headless
numpy
matplotlib
torch
torchvision
Pillow
requests
ultralytics-yolov5
